Massive commit; posix libc added, malloc(), IDTs, PIC programming, etc
This commit has been in the works and dormant over a few months. I got demotivated after attempts at IRQs and eventually pushed the code with it not working, hoping that someone could take a look at it and tell me what is wrong. That aside, the user-facing changes include malloc() for making applications and floating-point number support. If anyone would like to change the code or add new features such as IRQs, proper PIC, SMP support, etc. feel free to open a PR!