ps2_keyboard: Added modifier key support
With this commit, it is now possible to use uppercase
characters alongside lowercase ones and characters
otherwise only accessible through the shift/caps lock
layer. Here is a comprehensive list of all keys
now supported:

Layer 1:

1234567890-=\b
\tqwertyuiop[]\\
<CAPS>asdfghjkl;'<RETURN>
<SHIFT>zxcvbnm,./
<CTRL><ALT><SPACE>

Layer 2:

!@#$%^&*()_+\b
\tQWERTYUIOP{}|
<CAPS>ASDFGHJKL:"<RETURN>
<SHIFT>ZXCVBNM<>?
<CTRL><ALT><SPACE>

Notice the lack of Super/Meta/Mod4 support, as well as the lack of
right-side alt, shift, and ctrl support. This support is in the works,
but is not considered important because, realistically, nobody uses
them.